net/mlx5: fix HW steering incorrect job items offset
The HWS job memory contains header modify, header reformat, items.
And the memory layerout is also as this. A new job should start
the memory offset from the end of items. However, the job memory
start offset memory was from the end of header reformat, this
caused memory corruption.

This commit fixes the incorrect memory offset.

Fixes: 2c4126f8943b ("net/mlx5: add implicit port matching for HWS")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>