net/mlx5: reject hws counter age action on root table
In HW steering mode, if the counter action is on root table, using the
counter's offset to create flow rule is not supported (API is not
exposing this),
means it doesn't support counter in bulk usage which is the essential of
current HW steering's counter.

PMD, by default, doesn't create action on root table.

Needs to explicitly reject this and spit clear error message.
Since PMD only know group information when table creation, so add
validation of counter and age action in this.

Fixes: 2f9226a235b9 ("net/mlx5: add HWS direct AGE action support")
Fixes: 239d0c1e5762 ("net/mlx5: support HW steering count action")

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>