mlx5dr: Action, Add STE cleanup for single copy modify
Each action setter needs to assume that the STE given to him
is not memset to zero, for this we always clean the the STE
based on the type of action ctr/single/double.
On single modify header which is done inline we will clean
dw7 (containing the value) for ADD and SET while setting the
needed value, but we missed the set for COPY action.

Signed-off-by: Alex Vesker <valex@nvidia.com>